+++ components/layout.js
@@ -9,6 +9,7 @@ export default function Layout ({ children, instantSearch }) {
     <>
       <Head>
         <title>Video Browser</title>
+        <link rel='shortcut icon' href='/logo.svg' />
         <meta name='description' content="A browser for all the world's video" />
       </Head>
