+++ styles/Video.module.css
@@ -5,7 +5,6 @@
   cursor: pointer;
   font-size: 3em;
   font-weight: bold;
-  position: absolute;
   top: 2rem;
   left: 2rem;
 }
